Enable modulo matrix inverse + test

+++ content/graph/GeneralMatching.h
@@ -13,7 +13,7 @@
 #include "../numerical/MatrixInverse-mod.h"
 
 vector<pii> generalMatching(int N, vector<pii>& ed) {
-	vector<vector<ll>> mat(N, vector<ll>(N)), A;
+	vector<vector<int>> mat(N, vector<int>(N)), A;
 	for (pii pa : ed) {
 		int a = pa.first, b = pa.second, r = rand() % mod;
 		mat[a][b] = r, mat[b][a] = (mod - r) % mod;
@@ -23,7 +23,7 @@ vector<pii> generalMatching(int N, vector<pii>& ed) {
 	assert(r % 2 == 0);
 
 	if (M != N) do {
-		mat.resize(M, vector<ll>(M));
+		mat.resize(M, vector<int>(M));
 		rep(i,0,N) {
 			mat[i].resize(M);
 			rep(j,N,M) {
